requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.1.0
openai>=1.40.0
faiss-cpu>=1.8.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
pydantic>=2.7.0
python-dotenv>=1.0.1
fastapi>=0.112.0
uvicorn>=0.30.0
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
wikipedia>=1.4.0
rich>=13.7.1
pyyaml>=6.0.2
pytest>=8.2.0
langchain-openai>=0.3.33
langchain-chroma>=0.2.6
pypdf>=6.1.0
black>=25.9.0